Task 1.3. Buttons reassign actions without breakets.
