+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".MainActivity"
+    tools:context=".ui.MainActivity"
     android:background="@color/backgroundColor">
 
     <Space
